[ATOMIC-DESIGN-VUE-APP] Rename Form.vue & InputForm + scss files
