+++ src/assets/sass/main.scss
@@ -10,10 +10,10 @@
 
 // ATOMS
 @import './atoms/Button';
-@import './atoms/InputForm';
+@import './atoms/InputText';
 @import './atoms/ButtonIcon';
 
 // MOLECULES
 @import './pages/Buttons';
-@import './pages/Form';
+@import './pages/FormInputText';
 @import './pages/ButtonsIcons';
