+++ src/components/atoms/InputText.vue
